refactor(x/bank): move validate basic checks to msg server
