+++ x/bank/types/msgs.go
@@ -1,10 +1,7 @@
 package types
 
 import (
-	errorsmod "cosmossdk.io/errors"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 )
 
@@ -19,31 +16,8 @@ var (
 )
 
 // NewMsgSend - construct a msg to send coins from one account to another.
-//
-
 func NewMsgSend(fromAddr, toAddr sdk.AccAddress, amount sdk.Coins) *MsgSend {
 	return &MsgSend{FromAddress: fromAddr.String(), ToAddress: toAddr.String(), Amount: amount}
-}
-
-// ValidateBasic Implements Msg.
-func (msg MsgSend) ValidateBasic() error {
-	if _, err := sdk.AccAddressFromBech32(msg.FromAddress); err != nil {
-		return sdkerrors.ErrInvalidAddress.Wrapf("invalid from address: %s", err)
-	}
-
-	if _, err := sdk.AccAddressFromBech32(msg.ToAddress); err != nil {
-		return sdkerrors.ErrInvalidAddress.Wrapf("invalid to address: %s", err)
-	}
-
-	if !msg.Amount.IsValid() {
-		return errorsmod.Wrap(sdkerrors.ErrInvalidCoins, msg.Amount.String())
-	}
-
-	if !msg.Amount.IsAllPositive() {
-		return errorsmod.Wrap(sdkerrors.ErrInvalidCoins, msg.Amount.String())
-	}
-
-	return nil
 }
 
 // GetSignBytes Implements Msg.
@@ -62,26 +36,6 @@ func NewMsgMultiSend(in Input, out []Output) *MsgMultiSend {
 	return &MsgMultiSend{Inputs: []Input{in}, Outputs: out}
 }
 
-// ValidateBasic Implements Msg.
-func (msg MsgMultiSend) ValidateBasic() error {
-	// this just makes sure the input and all the outputs are properly formatted,
-	// not that they actually have the money inside
-
-	if len(msg.Inputs) == 0 {
-		return ErrNoInputs
-	}
-
-	if len(msg.Inputs) != 1 {
-		return ErrMultipleSenders
-	}
-
-	if len(msg.Outputs) == 0 {
-		return ErrNoOutputs
-	}
-
-	return ValidateInputOutputs(msg.Inputs[0], msg.Outputs)
-}
-
 // GetSignBytes Implements Msg.
 func (msg MsgMultiSend) GetSignBytes() []byte {
 	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&msg))
@@ -98,82 +52,6 @@ func (msg MsgMultiSend) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{addrs}
 }
 
-// ValidateBasic - validate transaction input
-func (in Input) ValidateBasic() error {
-	if _, err := sdk.AccAddressFromBech32(in.Address); err != nil {
-		return sdkerrors.ErrInvalidAddress.Wrapf("invalid input address: %s", err)
-	}
-
-	if !in.Coins.IsValid() {
-		return errorsmod.Wrap(sdkerrors.ErrInvalidCoins, in.Coins.String())
-	}
-
-	if !in.Coins.IsAllPositive() {
-		return errorsmod.Wrap(sdkerrors.ErrInvalidCoins, in.Coins.String())
-	}
-
-	return nil
-}
-
-// NewInput - create a transaction input, used with MsgMultiSend
-func NewInput(addr sdk.AccAddress, coins sdk.Coins) Input {
-	return Input{
-		Address: addr.String(),
-		Coins:   coins,
-	}
-}
-
-// ValidateBasic - validate transaction output
-func (out Output) ValidateBasic() error {
-	if _, err := sdk.AccAddressFromBech32(out.Address); err != nil {
-		return sdkerrors.ErrInvalidAddress.Wrapf("invalid output address: %s", err)
-	}
-
-	if !out.Coins.IsValid() {
-		return errorsmod.Wrap(sdkerrors.ErrInvalidCoins, out.Coins.String())
-	}
-
-	if !out.Coins.IsAllPositive() {
-		return errorsmod.Wrap(sdkerrors.ErrInvalidCoins, out.Coins.String())
-	}
-
-	return nil
-}
-
-// NewOutput - create a transaction output, used with MsgMultiSend
-func NewOutput(addr sdk.AccAddress, coins sdk.Coins) Output {
-	return Output{
-		Address: addr.String(),
-		Coins:   coins,
-	}
-}
-
-// ValidateInputOutputs validates that each respective input and output is
-// valid and that the sum of inputs is equal to the sum of outputs.
-func ValidateInputOutputs(input Input, outputs []Output) error {
-	var totalIn, totalOut sdk.Coins
-
-	if err := input.ValidateBasic(); err != nil {
-		return err
-	}
-	totalIn = input.Coins
-
-	for _, out := range outputs {
-		if err := out.ValidateBasic(); err != nil {
-			return err
-		}
-
-		totalOut = totalOut.Add(out.Coins...)
-	}
-
-	// make sure inputs and outputs match
-	if !totalIn.Equal(totalOut) {
-		return ErrInputOutputMismatch
-	}
-
-	return nil
-}
-
 // GetSigners returns the signer addresses that are expected to sign the result
 // of GetSignBytes.
 func (msg MsgUpdateParams) GetSigners() []sdk.AccAddress {
@@ -186,11 +64,6 @@ func (msg MsgUpdateParams) GetSigners() []sdk.AccAddress {
 func (msg MsgUpdateParams) GetSignBytes() []byte {
 	bz := ModuleCdc.MustMarshalJSON(&msg)
 	return sdk.MustSortJSON(bz)
-}
-
-// ValidateBasic performs basic MsgUpdateParams message validation.
-func (msg MsgUpdateParams) ValidateBasic() error {
-	return msg.Params.Validate()
 }
 
 // NewMsgSetSendEnabled Construct a message to set one or more SendEnabled entries.
@@ -211,34 +84,4 @@ func (msg MsgSetSendEnabled) GetSignBytes() []byte {
 func (msg MsgSetSendEnabled) GetSigners() []sdk.AccAddress {
 	addr, _ := sdk.AccAddressFromBech32(msg.Authority)
 	return []sdk.AccAddress{addr}
-}
-
-// ValidateBasic runs basic validation on this MsgSetSendEnabled.
-func (msg MsgSetSendEnabled) ValidateBasic() error {
-	if len(msg.Authority) > 0 {
-		if _, err := sdk.AccAddressFromBech32(msg.Authority); err != nil {
-			return sdkerrors.ErrInvalidAddress.Wrapf("invalid authority address: %s", err)
-		}
-	}
-
-	seen := map[string]bool{}
-	for _, se := range msg.SendEnabled {
-		if _, alreadySeen := seen[se.Denom]; alreadySeen {
-			return sdkerrors.ErrInvalidRequest.Wrapf("duplicate denom entries found for %q", se.Denom)
-		}
-
-		seen[se.Denom] = true
-
-		if err := se.Validate(); err != nil {
-			return sdkerrors.ErrInvalidRequest.Wrapf("invalid SendEnabled denom %q: %s", se.Denom, err)
-		}
-	}
-
-	for _, denom := range msg.UseDefaultFor {
-		if err := sdk.ValidateDenom(denom); err != nil {
-			return sdkerrors.ErrInvalidRequest.Wrapf("invalid UseDefaultFor denom %q: %s", denom, err)
-		}
-	}
-
-	return nil
 }
